Suggest video base on current play

+++ commands/command/plugin/music/Resume.js
@@ -1,22 +1,27 @@
-const { getPlayer, setPlay, isStop, setStop, getVC, getSong, destroyPlayer } = require('./Store')
+const { getPlayer, setPlay, isStop, setStop, getVC, getSong, destroyPlayer, AudioPlayer_exist } = require('./Store')
 const ytdl = require('./Ytdl-download')
 
-module.exports = (VoiceChannel) =>{
-    if(isStop(VoiceChannel.guild.id) === true){
-        console.log('Resume from stop')
-        setStop(VoiceChannel.guild.id, false)
-        const nextSong = getSong(VoiceChannel.guild.id)
-        if(nextSong !== false){
-            const stream = ytdl(nextSong)
-            getPlayer(VoiceChannel.guild.id).play(stream)
+module.exports = (VoiceChannel, message) =>{
+    if(AudioPlayer_exist(VoiceChannel.guild.id) !== false){
+        if(isStop(VoiceChannel.guild.id) === true){
+            console.log('Resume from stop')
+            setStop(VoiceChannel.guild.id, false)
+            const nextSong = getSong(VoiceChannel.guild.id)
+            if(nextSong !== false){
+                const stream = ytdl(nextSong)
+                getPlayer(VoiceChannel.guild.id).play(stream)
+            }
+            else{
+                destroyPlayer(VoiceChannel.guild.id)
+                getVC(VoiceChannel.guild.id).destroy()
+            }
         }
         else{
-            destroyPlayer(VoiceChannel.guild.id)
-            getVC(VoiceChannel.guild.id).destroy()
+            getPlayer(VoiceChannel.guild.id).unpause()
+            setPlay(VoiceChannel.guild.id, true)
         }
     }
     else{
-        getPlayer(VoiceChannel.guild.id).unpause()
-        setPlay(VoiceChannel.guild.id, true)
+        message.channel.send("Play music first");
     }
 }
